Introducing tests

+++ src/main/kotlin/com/barinek/bigstar/MetricsDataGateway.kt
@@ -2,7 +2,6 @@ package com.barinek.bigstar
 
 import org.springframework.jdbc.core.JdbcTemplate
 import org.springframework.stereotype.Repository
-import java.math.BigDecimal
 import javax.inject.Inject
 
 @Repository
@@ -19,9 +18,9 @@ class MetricsDataGateway {
                 { rs, rowNum ->
                     Metrics(
                             rs.getInt("number_of_accounts"),
-                            rs.getBigDecimal("total_account_value"))
+                            rs.getDouble("total_account_value"))
                 }
         )
-        return Metrics(10, BigDecimal.valueOf(100))
+        return Metrics(10, 100.00)
     }
 }
